I am trying to programmatically change the name of the family member on our member 1 shell
I added a list of family member names to the MyFamily java.  I am trying
to get the id of the family member button passed to the Member1 java so
that a different member name will show in the text view depending on the
id of the button pushed.  I haven't added the conditional statements to
the member1 java because I can't get this to work yet.

This is just to programmatically change the name of the family member in
the shell (member 1) activity.  We also need to figure out how to do
this for the text and the images.  It should be about the same once we
figure it out...