Added action bar menu on upper left
The action bar menu on the upper left now allows users to access "My
Famly", "My Profile" and "Logout".

I also added a conditional so that if you are logged in as "Grandma
Tata" the record button appears on both the "My Family" page and "My
Profile" page.  Previously, the record button appeared for "aa."